refactor(batch): decompose useBatchProcessor (2,281 -> 328 LOC, -86%)
Splits the largest hook in the codebase into a thin coordinator + 6 internal
hooks + 4 pure helpers under src/renderer/hooks/batch/internal/. Closes
section 4.3 of the Tier 4 backlog. Public API and behaviour unchanged; the
existing 165-test integration suite passes via the barrel.

What lives where:
- useBatchProcessor.ts        coordinator: shared refs, unmount cleanup, hook composition, public API
- batchFlushState.ts          AutoRunFlushState + claimFlushState() atomic-claim helper
- batchLoopSummary.ts         createLoopSummaryEntry pure builder
- batchFinalSummary.ts        buildFinalSummary pure builder ({ summary, details, isSuccess, statusText })
- batchProgressPoll.ts        createProgressPoll factory: { start, stop, restart }
- useBatchSelectors.ts        derived state + customPrompts handlers
- useBatchBroadcast.ts        broadcastAutoRunState + debounced updateBatchStateAndBroadcast + flushDebouncedUpdate
- useBatchControlActions.ts   stopBatchRun + pause/skip/resume/abort (share errorResolutionRefs + SET_STOPPING)
- useBatchKillAction.ts       killBatchRun + emergency flush via claimFlushState
- useBatchRunner.ts           startBatchRun: worktree setup, per-doc/per-task loop, stall detection, recount, history, Symphony, final-summary path

Notable design choices:
- claimFlushState extracted before the runner/kill split so both call sites
  share one atomic read-and-delete primitive. Removes the file's most subtle
  correctness invariant (kill vs natural-completion arbitration of the final
  history entry / endAutoRun stats / onComplete) from a "scattered across two
  files" failure mode.
- stopBatchRun merged into useBatchControlActions rather than its own file -
  shares errorResolutionRefs + SET_STOPPING with abortBatchOnError, separate
  hook would be ceremony.
- Coordinator owns all shared refs (stopRequestedRefs, errorResolutionRefs,
  autoRunFlushStateRefs, isMountedRef, sessionsRef, audioFeedback*Refs,
  updateBatchStateAndBroadcastRef) since each has 3+ readers/writers across
  the split. updateBatchStateAndBroadcastRef stays in the coordinator so the
  long-running async runner survives HMR re-renders across module boundaries.
- createProgressPoll.start() is async to preserve the original
  baseline-then-schedule sequence.

Behaviour delta: none. Same broadcast cadence (200 ms debounce), same stall
heuristic (3 consecutive no-progress runs), same progress-poll generation
guard (now hidden inside createProgressPoll rather than four leaked let
bindings), same kill-vs-natural-completion arbitration.

Tests:
- Existing 165-test integration suite passes unchanged via the hooks barrel.
- 63 new unit tests across 7 files covering each extracted unit.
- Full project suite: 27,603 passing / 108 skipped / 0 failing
  (was 27,468 baseline).
