+++ src/renderer/hooks/batch/useBatchProcessor.ts
@@ -1,4 +1,4 @@
-import { useCallback, useRef, useEffect, useMemo } from 'react';
+import { useCallback, useRef, useEffect } from 'react';
 import type {
 	BatchRunState,
 	BatchRunConfig,
@@ -9,34 +9,24 @@ import type {
 	AutoRunStats,
 	AgentError,
 } from '../../types';
-import {
-	getBadgeForTime,
-	getNextBadge,
-	formatTimeRemaining,
-} from '../../constants/conductorBadges';
-import { formatElapsedTime } from '../../../shared/formatters';
-import { gitService } from '../../services/git';
 // Extracted batch processing modules
 import { countUnfinishedTasks, uncheckAllTasks } from './batchUtils';
-import { useSessionDebounce } from './useSessionDebounce';
-import { DEFAULT_BATCH_STATE, type BatchAction } from './batchReducer';
-import { useBatchStore, selectHasAnyActiveBatch } from '../../stores/batchStore';
-import { useSessionStore, selectSessionById } from '../../stores/sessionStore';
-import { useSettingsStore } from '../../stores/settingsStore';
-import { notifyToast } from '../../stores/notificationStore';
+import { useBatchStore } from '../../stores/batchStore';
 import { useTimeTracking } from './useTimeTracking';
 import { useWorktreeManager } from './useWorktreeManager';
 import { useDocumentProcessor } from './useDocumentProcessor';
 import type { AgentSpawnErrorKind } from '../agent/useAgentExecution';
-import { logger } from '../../utils/logger';
-
-// Debounce delay for batch state updates (Quick Win 1)
-const BATCH_STATE_DEBOUNCE_MS = 200;
-const AUTO_RUN_PROGRESS_POLL_INTERVAL_MS = 20000;
-
-// Regex to match checked markdown checkboxes for reset-on-completion
-// Matches both [x] and [X] with various checkbox formats (standard and GitHub-style)
-// Note: countUnfinishedTasks, countCheckedTasks, uncheckAllTasks are now imported from ./batch/batchUtils
+// Decomposed internal hooks (see ./internal/)
+import type { BatchAction } from './batchReducer';
+import { type AutoRunFlushState } from './internal/batchFlushState';
+import { useBatchSelectors } from './internal/useBatchSelectors';
+import { useBatchBroadcast } from './internal/useBatchBroadcast';
+import {
+	useBatchControlActions,
+	type ErrorResolutionEntry,
+} from './internal/useBatchControlActions';
+import { useBatchKillAction } from './internal/useBatchKillAction';
+import { useBatchRunner } from './internal/useBatchRunner';
 
 export interface BatchCompleteInfo {
 	sessionId: string;
@@ -126,85 +116,6 @@ interface UseBatchProcessorReturn {
 	abortBatchOnError: (sessionId: string) => void;
 }
 
-type ErrorResolutionAction = 'resume' | 'skip-document' | 'abort';
-
-interface ErrorResolutionEntry {
-	promise: Promise<ErrorResolutionAction>;
-	resolve: (action: ErrorResolutionAction) => void;
-}
-
-/**
- * Create a loop summary history entry
- */
-interface LoopSummaryParams {
-	loopIteration: number;
-	loopTasksCompleted: number;
-	loopStartTime: number;
-	loopTotalInputTokens: number;
-	loopTotalOutputTokens: number;
-	loopTotalCost: number;
-	sessionCwd: string;
-	sessionId: string;
-	isFinal: boolean;
-	exitReason?: string;
-}
-
-function createLoopSummaryEntry(params: LoopSummaryParams): Omit<HistoryEntry, 'id'> {
-	const {
-		loopIteration,
-		loopTasksCompleted,
-		loopStartTime,
-		loopTotalInputTokens,
-		loopTotalOutputTokens,
-		loopTotalCost,
-		sessionCwd,
-		sessionId,
-		isFinal,
-		exitReason,
-	} = params;
-
-	const loopElapsedMs = Date.now() - loopStartTime;
-	const loopNumber = loopIteration + 1;
-	const summaryPrefix = isFinal ? `Loop ${loopNumber} (final)` : `Loop ${loopNumber}`;
-	const loopSummary = `${summaryPrefix} completed: ${loopTasksCompleted} task${loopTasksCompleted !== 1 ? 's' : ''} accomplished`;
-
-	const loopDetails = [
-		`**${summaryPrefix} Summary**`,
-		'',
-		`- **Tasks Accomplished:** ${loopTasksCompleted}`,
-		`- **Duration:** ${formatElapsedTime(loopElapsedMs)}`,
-		loopTotalInputTokens > 0 || loopTotalOutputTokens > 0
-			? `- **Tokens:** ${(loopTotalInputTokens + loopTotalOutputTokens).toLocaleString()} (${loopTotalInputTokens.toLocaleString()} in / ${loopTotalOutputTokens.toLocaleString()} out)`
-			: '',
-		loopTotalCost > 0 ? `- **Cost:** $${loopTotalCost.toFixed(4)}` : '',
-		exitReason ? `- **Exit Reason:** ${exitReason}` : '',
-	]
-		.filter((line) => line !== '')
-		.join('\n');
-
-	return {
-		type: 'AUTO',
-		timestamp: Date.now(),
-		summary: loopSummary,
-		fullResponse: loopDetails,
-		projectPath: sessionCwd,
-		sessionId: sessionId,
-		success: true,
-		elapsedTimeMs: loopElapsedMs,
-		usageStats:
-			loopTotalInputTokens > 0 || loopTotalOutputTokens > 0
-				? {
-						inputTokens: loopTotalInputTokens,
-						outputTokens: loopTotalOutputTokens,
-						cacheReadInputTokens: 0,
-						cacheCreationInputTokens: 0,
-						totalCostUsd: loopTotalCost,
-						contextWindow: 0,
-					}
-				: undefined,
-	};
-}
-
 // Re-export utility functions for backwards compatibility
 // (countUnfinishedTasks and uncheckAllTasks are imported from ./batch/batchUtils)
 export { countUnfinishedTasks, uncheckAllTasks };
@@ -231,17 +142,22 @@ export function useBatchProcessor({
 	autoRunStats,
 	onProcessQueueAfterCompletion,
 }: UseBatchProcessorProps): UseBatchProcessorReturn {
-	// Batch states per session — lives in batchStore, read reactively for re-renders
-	const batchRunStates = useBatchStore((s) => s.batchRunStates);
+	// Reactive selectors over the batch store
+	const {
+		batchRunStates,
+		getBatchState,
+		hasAnyActiveBatch,
+		activeBatchSessionIds,
+		stoppingBatchSessionIds,
+		customPrompts,
+		setCustomPrompt,
+	} = useBatchSelectors();
 
 	// Dispatch batch actions through the store. The store applies batchReducer
 	// synchronously, eliminating the need for manual ref syncing.
 	const dispatch = useCallback((action: BatchAction) => {
 		useBatchStore.getState().dispatchBatch(action);
 	}, []);
-
-	// Custom prompts per session — lives in batchStore
-	const customPrompts = useBatchStore((s) => s.customPrompts);
 
 	// Refs for tracking stop requests per session
 	const stopRequestedRefs = useRef<Record<string, boolean>>({});
@@ -268,22 +184,7 @@ export function useBatchProcessor({
 	// responsible for writing the final history + endAutoRun. This guards against the
 	// case where killBatchRun calls timeTracking.stopTracking (which zeros the tracker)
 	// before the loop's cleanup reads it, resulting in a 0ms duration being recorded.
-	const autoRunFlushStateRefs = useRef<
-		Record<
-			string,
-			{
-				statsAutoRunId: string | null;
-				sessionName: string;
-				projectPath: string;
-				getCompletedTasks: () => number;
-				getTotalTasks: () => number;
-				getInputTokens: () => number;
-				getOutputTokens: () => number;
-				getTotalCost: () => number;
-				getDocumentsProcessed: () => number;
-			}
-		>
-	>({});
+	const autoRunFlushStateRefs = useRef<Record<string, AutoRunFlushState>>({});
 
 	// Track whether the component is still mounted to prevent state updates after unmount
 	const isMountedRef = useRef(false);
@@ -311,138 +212,24 @@ export function useBatchProcessor({
 		};
 	}, []);
 
-	/**
-	 * Broadcast Auto Run state to web interface immediately (synchronously).
-	 * This replaces the previous useEffect-based approach to ensure mobile clients
-	 * receive state updates without waiting for React's render cycle.
-	 */
-	const broadcastAutoRunState = useCallback((sessionId: string, state: BatchRunState | null) => {
-		if (
-			state &&
-			(state.isRunning || state.completedTasks > 0 || state.completedTasksAcrossAllDocs > 0)
-		) {
-			window.maestro.web.broadcastAutoRunState(sessionId, {
-				isRunning: state.isRunning,
-				totalTasks: state.totalTasks,
-				completedTasks: state.completedTasks,
-				currentTaskIndex: state.currentTaskIndex,
-				isStopping: state.isStopping,
-				// Multi-document progress fields
-				totalDocuments: state.documents?.length ?? 0,
-				currentDocumentIndex: state.currentDocumentIndex,
-				totalTasksAcrossAllDocs: state.totalTasksAcrossAllDocs,
-				completedTasksAcrossAllDocs: state.completedTasksAcrossAllDocs,
-				// Error pause fields — surfaced to web/mobile so they can show recovery UI
-				errorPaused: state.errorPaused,
-				errorMessage: state.error?.message,
-				errorType: state.error?.type,
-				errorRecoverable: state.error?.recoverable,
-				errorDocumentIndex: state.errorDocumentIndex,
-				errorTaskDescription: state.errorTaskDescription,
-			});
-		} else {
-			// When not running and no completed tasks, broadcast null to clear the state
-			window.maestro.web.broadcastAutoRunState(sessionId, null);
-		}
-	}, []);
+	// Web/mobile bridge: synchronous broadcast + debounced state-update wrapper
+	const { broadcastAutoRunState, updateBatchStateAndBroadcast, flushDebouncedUpdate } =
+		useBatchBroadcast({ dispatch });
 
-	// Use extracted debounce hook for batch state updates (replaces manual debounce logic)
-	const { scheduleUpdate: _scheduleDebouncedUpdate, flushUpdate: flushDebouncedUpdate } =
-		useSessionDebounce<Record<string, BatchRunState>>({
-			delayMs: BATCH_STATE_DEBOUNCE_MS,
-			onUpdate: useCallback(
-				(
-					sessionId: string,
-					updater: (prev: Record<string, BatchRunState>) => Record<string, BatchRunState>
-				) => {
-					// Apply the updater and get the new state for broadcasting
-					// Note: We use a ref to capture the new state since dispatch doesn't return it
-					let newStateForSession: BatchRunState | null = null;
-
-					try {
-						// For reducer, we need to convert the updater to an action
-						// Since the updater pattern doesn't map directly to actions, we wrap it
-						// by reading current state and computing the new state
-						const currentState = useBatchStore.getState().batchRunStates;
-						const newState = updater(currentState);
-						newStateForSession = newState[sessionId] || null;
-
-						// Dispatch UPDATE_PROGRESS with the computed changes
-						// For complex state changes, we extract the session's new state and dispatch appropriately
-						if (newStateForSession) {
-							const prevSessionState = currentState[sessionId] || DEFAULT_BATCH_STATE;
-
-							// Dispatch UPDATE_PROGRESS with any changed fields
-							dispatch({
-								type: 'UPDATE_PROGRESS',
-								sessionId,
-								payload: {
-									currentDocumentIndex:
-										newStateForSession.currentDocumentIndex !==
-										prevSessionState.currentDocumentIndex
-											? newStateForSession.currentDocumentIndex
-											: undefined,
-									currentDocTasksTotal:
-										newStateForSession.currentDocTasksTotal !==
-										prevSessionState.currentDocTasksTotal
-											? newStateForSession.currentDocTasksTotal
-											: undefined,
-									currentDocTasksCompleted:
-										newStateForSession.currentDocTasksCompleted !==
-										prevSessionState.currentDocTasksCompleted
-											? newStateForSession.currentDocTasksCompleted
-											: undefined,
-									totalTasksAcrossAllDocs:
-										newStateForSession.totalTasksAcrossAllDocs !==
-										prevSessionState.totalTasksAcrossAllDocs
-											? newStateForSession.totalTasksAcrossAllDocs
-											: undefined,
-									completedTasksAcrossAllDocs:
-										newStateForSession.completedTasksAcrossAllDocs !==
-										prevSessionState.completedTasksAcrossAllDocs
-											? newStateForSession.completedTasksAcrossAllDocs
-											: undefined,
-									totalTasks:
-										newStateForSession.totalTasks !== prevSessionState.totalTasks
-											? newStateForSession.totalTasks
-											: undefined,
-									completedTasks:
-										newStateForSession.completedTasks !== prevSessionState.completedTasks
-											? newStateForSession.completedTasks
-											: undefined,
-									currentTaskIndex:
-										newStateForSession.currentTaskIndex !== prevSessionState.currentTaskIndex
-											? newStateForSession.currentTaskIndex
-											: undefined,
-									sessionIds:
-										newStateForSession.sessionIds !== prevSessionState.sessionIds
-											? newStateForSession.sessionIds
-											: undefined,
-									accumulatedElapsedMs:
-										newStateForSession.accumulatedElapsedMs !==
-										prevSessionState.accumulatedElapsedMs
-											? newStateForSession.accumulatedElapsedMs
-											: undefined,
-									lastActiveTimestamp:
-										newStateForSession.lastActiveTimestamp !== prevSessionState.lastActiveTimestamp
-											? newStateForSession.lastActiveTimestamp
-											: undefined,
-									loopIteration:
-										newStateForSession.loopIteration !== prevSessionState.loopIteration
-											? newStateForSession.loopIteration
-											: undefined,
-								},
-							});
-						}
-
-						broadcastAutoRunState(sessionId, newStateForSession);
-					} catch (error) {
-						logger.error('[BatchProcessor:onUpdate] ERROR in debounce callback:', undefined, error);
-					}
-				},
-				[broadcastAutoRunState]
-			),
-		});
+	// External lifecycle controls (stop + pause/skip/resume/abort)
+	const {
+		stopBatchRun,
+		pauseBatchOnError,
+		skipCurrentDocument,
+		resumeAfterError,
+		abortBatchOnError,
+	} = useBatchControlActions({
+		broadcastAutoRunState,
+		dispatch,
+		errorResolutionRefs,
+		stopRequestedRefs,
+		isMountedRef,
+	});
 
 	// Use extracted time tracking hook (replaces manual visibility-based time tracking)
 	const timeTracking = useTimeTracking({
@@ -467,1799 +254,58 @@ export function useBatchProcessor({
 		),
 	});
 
+	// Force-kill action with kill-vs-natural-completion arbitration.
+	// Must follow `useTimeTracking` because it captures elapsed time from the
+	// live tracker before tearing it down.
+	const { killBatchRun } = useBatchKillAction({
+		broadcastAutoRunState,
+		flushDebouncedUpdate,
+		dispatch,
+		timeTracking,
+		autoRunFlushStateRefs,
+		errorResolutionRefs,
+		stopRequestedRefs,
+		isMountedRef,
+		onAddHistoryEntry,
+		onComplete,
+	});
+
 	// Use extracted worktree manager hook for git worktree operations
 	const worktreeManager = useWorktreeManager();
 
 	// Use extracted document processor hook for document processing
 	const documentProcessor = useDocumentProcessor();
 
-	// Helper to get batch state for a session
-	// Note: This reads from React state (not the ref) because consumers need React
-	// to trigger re-renders when state changes. The ref is used internally for
-	// synchronous access in debounced callbacks.
-	const getBatchState = useCallback(
-		(sessionId: string): BatchRunState => {
-			return batchRunStates[sessionId] || DEFAULT_BATCH_STATE;
-		},
-		[batchRunStates]
-	);
-
-	// Boolean selector is stable with Object.is comparison
-	const hasAnyActiveBatch = useBatchStore(selectHasAnyActiveBatch);
-
-	// Array selectors use useMemo to avoid infinite re-renders
-	// (Zustand's Object.is comparison treats new arrays as changed → re-render loop)
-	const activeBatchSessionIds = useMemo(
-		() =>
-			Object.entries(batchRunStates)
-				.filter(([, state]) => state.isRunning && !state.errorPaused)
-				.map(([sessionId]) => sessionId),
-		[batchRunStates]
-	);
-	const stoppingBatchSessionIds = useMemo(
-		() =>
-			Object.entries(batchRunStates)
-				.filter(([, state]) => state.isRunning && state.isStopping)
-				.map(([sessionId]) => sessionId),
-		[batchRunStates]
-	);
-
-	// Set custom prompt for a session (delegates to store)
-	const setCustomPrompt = useCallback((sessionId: string, prompt: string) => {
-		useBatchStore.getState().setCustomPrompt(sessionId, prompt);
-	}, []);
-
-	/**
-	 * Update batch state AND broadcast to web interface with debouncing.
-	 * This wrapper uses the extracted useSessionDebounce hook to batch rapid-fire
-	 * state updates and reduce React re-renders during intensive task processing.
-	 *
-	 * Critical updates (isRunning changes, errors) are processed immediately,
-	 * while progress updates are debounced by BATCH_STATE_DEBOUNCE_MS.
-	 */
-	const updateBatchStateAndBroadcast = useCallback(
-		(
-			sessionId: string,
-			updater: (prev: Record<string, BatchRunState>) => Record<string, BatchRunState>,
-			immediate: boolean = false
-		) => {
-			_scheduleDebouncedUpdate(sessionId, updater, immediate);
-		},
-		[_scheduleDebouncedUpdate]
-	);
-
-	// Update ref to always have latest updateBatchStateAndBroadcast (fixes HMR stale closure)
+	// Update ref to always have latest updateBatchStateAndBroadcast (fixes HMR stale closure
+	// in long-running async loops; safe across module boundaries because Vite invalidates
+	// per-module — keeping the ref in the coordinator is intentional).
 	updateBatchStateAndBroadcastRef.current = updateBatchStateAndBroadcast;
 
-	// Use readDocAndCountTasks from the extracted documentProcessor hook
-	// This replaces the previous inline helper function
-	const readDocAndCountTasks = documentProcessor.readDocAndCountTasks;
-
-	/**
-	 * Start a batch processing run for a specific session with multi-document support.
-	 * Note: sessionId and folderPath can belong to different sessions when running
-	 * in a worktree — the parent session owns the Auto Run documents (folderPath)
-	 * while the worktree agent (sessionId) executes the tasks.
-	 */
-	const startBatchRun = useCallback(
-		async (sessionId: string, config: BatchRunConfig, folderPath: string) => {
-			// Check global Auto Run kill switch
-			if (useSettingsStore.getState().autoRunDisabled) {
-				window.maestro.logger.log(
-					'warn',
-					'Auto Run is disabled via autoRunDisabled setting',
-					'BatchProcessor',
-					{ sessionId }
-				);
-				notifyToast({
-					type: 'warning',
-					title: 'Auto Run Disabled',
-					message: 'Auto Run is disabled. Enable it in Settings to use this feature.',
-				});
-				return;
-			}
-
-			window.maestro.logger.log('info', 'startBatchRun called', 'BatchProcessor', {
-				sessionId,
-				folderPath,
-				documentsCount: config.documents.length,
-				worktreeEnabled: config.worktree?.enabled,
-			});
-
-			// Use sessionsRef first, then fall back to Zustand store for sessions just created
-			// (sessionsRef updates on React re-render, but Zustand store updates synchronously)
-			const session =
-				sessionsRef.current.find((s) => s.id === sessionId) ||
-				selectSessionById(sessionId)(useSessionStore.getState());
-			if (!session) {
-				const worktreeInfo = config.worktreeTarget
-					? ` (worktree mode: ${config.worktreeTarget.mode}, path: ${
-							config.worktreeTarget.mode === 'existing-closed'
-								? config.worktreeTarget.worktreePath
-								: config.worktreeTarget.mode === 'create-new'
-									? config.worktreeTarget.newBranchName
-									: config.worktreeTarget.sessionId
-						})`
-					: '';
-				window.maestro.logger.log(
-					'error',
-					`Session not found for batch processing${worktreeInfo}`,
-					'BatchProcessor',
-					{
-						sessionId,
-						worktreeTargetMode: config.worktreeTarget?.mode,
-						availableSessionIds: sessionsRef.current.map((s) => s.id),
-					}
-				);
-				return;
-			}
-
-			const { documents, prompt, loopEnabled, maxLoops, worktree } = config;
-
-			if (documents.length === 0) {
-				window.maestro.logger.log(
-					'warn',
-					'No documents provided for batch processing',
-					'BatchProcessor',
-					{ sessionId }
-				);
-				return;
-			}
-
-			// Track batch start time for completion notification
-			const batchStartTime = Date.now();
-
-			// Initialize visibility-based time tracking for this session using the extracted hook
-			timeTracking.startTracking(sessionId);
-
-			// Reset stop flag for this session
-			stopRequestedRefs.current[sessionId] = false;
-			delete errorResolutionRefs.current[sessionId];
-
-			// Set up worktree if enabled using extracted hook.
-			// Note: sshRemoteId is only set after AI agent spawns. For terminal-only SSH sessions,
-			// we must fall back to sessionSshRemoteConfig.remoteId. See CLAUDE.md "SSH Remote Sessions".
-			const sshRemoteId =
-				session.sshRemoteId || session.sessionSshRemoteConfig?.remoteId || undefined;
-
-			let effectiveCwd: string;
-			let worktreeActive: boolean;
-			let worktreePath: string | undefined;
-			let worktreeBranch: string | undefined;
-
-			if (config.worktreeTarget) {
-				// Worktree dispatch was already handled by useAutoRunHandlers
-				// (spawnWorktreeAgentAndDispatch created the worktree and session).
-				// Skip setupWorktree — calling it again would fail because the session's
-				// CWD is already a worktree, not the main repo, causing a
-				// "belongs to a different repository" false positive.
-				effectiveCwd = session.cwd;
-				worktreeActive = true;
-				worktreePath = session.cwd;
-				worktreeBranch = session.worktreeBranch || config.worktree?.branchName;
-			} else {
-				// Normal path: set up worktree from scratch if config.worktree is enabled
-				const worktreeWithSsh = worktree ? { ...worktree, sshRemoteId } : undefined;
-				const worktreeResult = await worktreeManager.setupWorktree(session.cwd, worktreeWithSsh);
-				if (!worktreeResult.success) {
-					window.maestro.logger.log('error', 'Worktree setup failed', 'BatchProcessor', {
-						sessionId,
-						error: worktreeResult.error,
-					});
-					return;
-				}
-				effectiveCwd = worktreeResult.effectiveCwd;
-				worktreeActive = worktreeResult.worktreeActive;
-				worktreePath = worktreeResult.worktreePath;
-				worktreeBranch = worktreeResult.worktreeBranch;
-			}
-
-			// Get git branch for template variable substitution
-			let gitBranch: string | undefined;
-			if (session.isGitRepo) {
-				try {
-					const status = await gitService.getStatus(effectiveCwd);
-					gitBranch = status.branch;
-				} catch {
-					// Ignore git errors - branch will be empty string
-				}
-			}
-
-			// Find group name for this session (sessions have groupId, groups have id)
-			const sessionGroup = session.groupId ? groups.find((g) => g.id === session.groupId) : null;
-			const groupName = sessionGroup?.name;
-
-			// Calculate initial total tasks across all documents (checked + unchecked)
-			let initialTotalTasks = 0;
-			let initialCheckedTasks = 0;
-			for (const doc of documents) {
-				const { taskCount, checkedCount } = await readDocAndCountTasks(
-					folderPath,
-					doc.filename,
-					sshRemoteId
-				);
-				initialTotalTasks += taskCount + checkedCount;
-				initialCheckedTasks += checkedCount;
-			}
-			// Track unchecked count for the "no tasks" early exit check
-			const initialUncheckedTasks = initialTotalTasks - initialCheckedTasks;
-
-			if (initialUncheckedTasks === 0) {
-				window.maestro.logger.log(
-					'warn',
-					'No unchecked tasks found across all documents',
-					'BatchProcessor',
-					{ sessionId }
-				);
-				return;
-			}
-
-			// Initialize batch run state using START_BATCH action directly
-			// (not updateBatchStateAndBroadcast which only supports UPDATE_PROGRESS)
-			const lockedDocuments = documents.map((d) => d.filename);
-			dispatch({
-				type: 'START_BATCH',
-				sessionId,
-				payload: {
-					documents: documents.map((d) => d.filename),
-					lockedDocuments,
-					totalTasksAcrossAllDocs: initialTotalTasks,
-					completedTasksAcrossAllDocs: initialCheckedTasks,
-					loopEnabled,
-					maxLoops,
-					folderPath,
-					worktreeActive,
-					worktreePath,
-					worktreeBranch,
-					customPrompt: prompt !== '' ? prompt : undefined,
-					startTime: batchStartTime,
-					// Time tracking
-					cumulativeTaskTimeMs: 0, // Sum of actual task durations (most accurate)
-					accumulatedElapsedMs: 0, // Visibility-based time (excludes sleep/suspend)
-					lastActiveTimestamp: batchStartTime,
-				},
-			});
-			// Broadcast state change
-			broadcastAutoRunState(sessionId, {
-				isRunning: true,
-				isStopping: false,
-				documents: documents.map((d) => d.filename),
-				lockedDocuments,
-				currentDocumentIndex: 0,
-				currentDocTasksTotal: 0,
-				currentDocTasksCompleted: 0,
-				totalTasksAcrossAllDocs: initialTotalTasks,
-				completedTasksAcrossAllDocs: 0,
-				loopEnabled,
-				loopIteration: 0,
-				maxLoops,
-				folderPath,
-				worktreeActive,
-				worktreePath,
-				worktreeBranch,
-				totalTasks: initialTotalTasks,
-				completedTasks: 0,
-				currentTaskIndex: 0,
-				originalContent: '',
-				customPrompt: prompt !== '' ? prompt : undefined,
-				sessionIds: [],
-				startTime: batchStartTime,
-				accumulatedElapsedMs: 0,
-				lastActiveTimestamp: batchStartTime,
-			});
-
-			// AUTORUN LOG: Start
-			window.maestro.logger.autorun(`Auto Run started`, session.name, {
-				documents: documents.map((d) => d.filename),
-				totalTasks: initialTotalTasks,
-				loopEnabled,
-				maxLoops: maxLoops ?? 'unlimited',
-			});
-
-			// Notify user that Auto Run has started
-			notifyToast({
-				type: 'info',
-				title: 'Auto Run Started',
-				message: `${initialTotalTasks} ${initialTotalTasks === 1 ? 'task' : 'tasks'} across ${documents.length} ${documents.length === 1 ? 'document' : 'documents'}`,
-				project: session.name,
-				sessionId,
-			});
-
-			// Add initial history entry when using worktree
-			if (worktreeActive && worktreePath && worktreeBranch) {
-				const worktreeStartSummary = `Auto Run started in worktree`;
-				const worktreeStartDetails = [
-					`**Worktree Auto Run Started**`,
-					``,
-					`- **Branch:** \`${worktreeBranch}\``,
-					`- **Worktree Path:** \`${worktreePath}\``,
-					`- **Main Repo:** \`${session.cwd}\``,
-					`- **Documents:** ${documents.map((d) => d.filename).join(', ')}`,
-					`- **Total Tasks:** ${initialTotalTasks}`,
-					loopEnabled ? `- **Loop Mode:** Enabled${maxLoops ? ` (max ${maxLoops})` : ''}` : '',
-				]
-					.filter((line) => line !== '')
-					.join('\n');
-
-				onAddHistoryEntry({
-					type: 'AUTO',
-					timestamp: Date.now(),
-					summary: worktreeStartSummary,
-					fullResponse: worktreeStartDetails,
-					projectPath: effectiveCwd,
-					sessionId: sessionId,
-					success: true,
-				});
-			}
-
-			// Store custom prompt for persistence
-			useBatchStore.getState().setCustomPrompt(sessionId, prompt);
-
-			// State machine: INITIALIZING -> RUNNING (initialization complete)
-			dispatch({ type: 'SET_RUNNING', sessionId });
-
-			// Prevent system sleep while Auto Run is active
-			window.maestro.power.addReason(`autorun:${sessionId}`);
-
-			// Start stats tracking for this Auto Run session
-			let statsAutoRunId: string | null = null;
-			try {
-				statsAutoRunId = await window.maestro.stats.startAutoRun({
-					sessionId: sessionId,
-					agentType: session.toolType,
-					documentPath: documents.map((d) => d.filename).join(', '),
-					startTime: batchStartTime,
-					tasksTotal: initialTotalTasks,
-					projectPath: session.cwd,
-				});
-			} catch (statsError) {
-				// Don't fail the batch if stats tracking fails
-				logger.warn('[BatchProcessor] Failed to start stats tracking:', undefined, statsError);
-			}
-
-			// Collect Claude session IDs and track completion
-			const agentSessionIds: string[] = [];
-			let totalCompletedTasks = 0;
-			let loopIteration = 0;
-
-			// Register this Auto Run for emergency stats/history flush on force-kill.
-			// Populated even if startAutoRun failed (statsAutoRunId null) so killBatchRun can
-			// still write a history entry with the elapsed time the user actually spent.
-			autoRunFlushStateRefs.current[sessionId] = {
-				statsAutoRunId,
-				sessionName: session.name || session.cwd.split('/').pop() || 'Unknown',
-				projectPath: session.cwd,
-				getCompletedTasks: () => totalCompletedTasks,
-				getTotalTasks: () => initialTotalTasks,
-				getInputTokens: () => totalInputTokens,
-				getOutputTokens: () => totalOutputTokens,
-				getTotalCost: () => totalCost,
-				getDocumentsProcessed: () => documents.length,
-			};
-
-			// Per-loop tracking for loop summary
-			let loopStartTime = Date.now();
-			let loopTasksCompleted = 0;
-			let loopTotalInputTokens = 0;
-			let loopTotalOutputTokens = 0;
-			let loopTotalCost = 0;
-
-			// Cumulative tracking for final Auto Run summary (across all loops)
-			let totalInputTokens = 0;
-			let totalOutputTokens = 0;
-			let totalCost = 0;
-
-			// Track consecutive runs with no task-level progress (nothing checked off, no tasks
-			// added or removed). Content-level comparison is unreliable because the agent can
-			// mutate the doc (append addenda/explanation text) without doing actual work, which
-			// would reset a content-based counter and hide the stall indefinitely.
-			// Note: This counter is reset per-document, so stalling one document doesn't affect others
-			let consecutiveNoChangeCount = 0;
-			const MAX_CONSECUTIVE_NO_CHANGES = 3; // Skip document after 3 consecutive runs with no task-level progress
-
-			// Track stalled documents (document filename -> stall reason)
-			const stalledDocuments: Map<string, string> = new Map();
-
-			// Track working copies for reset-on-completion documents (original filename -> working copy path)
-			// Working copies are stored in /Runs/ and serve as audit logs
-			const workingCopies: Map<string, string> = new Map();
-
-			// Helper to add final loop summary (defined here so it has access to tracking vars)
-			const addFinalLoopSummary = (exitReason: string) => {
-				// AUTORUN LOG: Exit
-				window.maestro.logger.autorun(`Auto Run exiting: ${exitReason}`, session.name, {
-					reason: exitReason,
-					totalTasksCompleted: totalCompletedTasks,
-					loopsCompleted: loopIteration + 1,
-				});
-
-				if (loopEnabled && (loopTasksCompleted > 0 || loopIteration > 0)) {
-					onAddHistoryEntry(
-						createLoopSummaryEntry({
-							loopIteration,
-							loopTasksCompleted,
-							loopStartTime,
-							loopTotalInputTokens,
-							loopTotalOutputTokens,
-							loopTotalCost,
-							sessionCwd: session.cwd,
-							sessionId,
-							isFinal: true,
-							exitReason,
-						})
-					);
-				}
-			};
-
-			// Main processing loop (handles loop mode)
-			while (true) {
-				// Check for stop request
-				if (stopRequestedRefs.current[sessionId]) {
-					addFinalLoopSummary('Stopped by user');
-					break;
-				}
-
-				// Track if any tasks were processed in this iteration
-				let anyTasksProcessedThisIteration = false;
-
-				// Process each document in order
-				for (let docIndex = 0; docIndex < documents.length; docIndex++) {
-					// Check for stop request before each document
-					if (stopRequestedRefs.current[sessionId]) {
-						break;
-					}
-
-					const docEntry = documents[docIndex];
-
-					// Read document and count tasks
-					let {
-						taskCount: remainingTasks,
-						content: docContent,
-						checkedCount: docCheckedCount,
-					} = await readDocAndCountTasks(folderPath, docEntry.filename, sshRemoteId);
-					let docTasksTotal = remainingTasks;
-
-					// Handle documents with no unchecked tasks
-					if (remainingTasks === 0) {
-						// For reset-on-completion documents, check if there are checked tasks that need resetting
-						if (docEntry.resetOnCompletion && loopEnabled) {
-							// Use docCheckedCount from readDocAndCountTasks instead of calling countCheckedTasks again
-							if (docCheckedCount > 0) {
-								const resetContent = uncheckAllTasks(docContent);
-								await window.maestro.autorun.writeDoc(
-									folderPath,
-									docEntry.filename + '.md',
-									resetContent,
-									sshRemoteId
-								);
-								// Update task count in state
-								const resetTaskCount = countUnfinishedTasks(resetContent);
-								updateBatchStateAndBroadcastRef.current!(sessionId, (prev) => ({
-									...prev,
-									[sessionId]: {
-										...prev[sessionId],
-										totalTasksAcrossAllDocs:
-											prev[sessionId].totalTasksAcrossAllDocs + resetTaskCount,
-										totalTasks: prev[sessionId].totalTasks + resetTaskCount,
-									},
-								}));
-							}
-						}
-						continue;
-					}
-
-					// Reset stall detection counter for each new document
-					consecutiveNoChangeCount = 0;
-
-					// The actual filename to process (may be working copy for reset-on-completion docs)
-					let effectiveFilename = docEntry.filename;
-
-					// Create working copy for reset-on-completion documents
-					// Working copies are stored in /Runs/ and the original is never modified
-					if (docEntry.resetOnCompletion) {
-						try {
-							const { workingCopyPath } = await window.maestro.autorun.createWorkingCopy(
-								folderPath,
-								docEntry.filename,
-								loopIteration + 1, // 1-indexed loop number
-								sshRemoteId
-							);
-							workingCopies.set(docEntry.filename, workingCopyPath);
-							effectiveFilename = workingCopyPath;
-
-							// Re-read the working copy for task counting
-							const workingCopyResult = await readDocAndCountTasks(
-								folderPath,
-								effectiveFilename,
-								sshRemoteId
-							);
-							remainingTasks = workingCopyResult.taskCount;
-							docContent = workingCopyResult.content;
-							docCheckedCount = workingCopyResult.checkedCount;
-							docTasksTotal = remainingTasks;
-						} catch (err) {
-							logger.error(
-								`[BatchProcessor] Failed to create working copy for ${docEntry.filename}:`,
-								undefined,
-								err
-							);
-							// Continue with original document as fallback
-						}
-					}
-
-					// AUTORUN LOG: Document processing
-					window.maestro.logger.autorun(`Processing document: ${docEntry.filename}`, session.name, {
-						document: docEntry.filename,
-						tasksRemaining: remainingTasks,
-						loopNumber: loopIteration + 1,
-					});
-
-					// Update state to show current document
-					updateBatchStateAndBroadcastRef.current!(sessionId, (prev) => ({
-						...prev,
-						[sessionId]: {
-							...prev[sessionId],
-							currentDocumentIndex: docIndex,
-							currentDocTasksTotal: docTasksTotal,
-							currentDocTasksCompleted: 0,
-						},
-					}));
-
-					let docTasksCompleted = 0;
-					let skipCurrentDocumentAfterError = false;
-
-					// Process tasks in this document until none remain
-					while (remainingTasks > 0) {
-						// Check for stop request before each task
-						if (stopRequestedRefs.current[sessionId]) {
-							break;
-						}
-
-						// Pause processing until the user resolves the error state
-						const errorResolution = errorResolutionRefs.current[sessionId];
-						if (errorResolution) {
-							const action = await errorResolution.promise;
-							delete errorResolutionRefs.current[sessionId];
-
-							if (action === 'abort') {
-								stopRequestedRefs.current[sessionId] = true;
-								break;
-							}
-
-							if (action === 'skip-document') {
-								skipCurrentDocumentAfterError = true;
-								break;
-							}
-						}
-
-						// Use extracted document processor hook for task processing
-						// This handles: template substitution, document expansion, agent spawning,
-						// session registration, re-reading document, and synopsis generation
-
-						// Poll only the currently-processing document. Other documents in the
-						// playbook can't change during this task — the agent is working on
-						// docEntry — so snapshot their counts once and reuse them across ticks.
-						let progressPollActive = true;
-						let progressPollInFlight = false;
-						let progressPollGeneration = 0;
-						let progressPollTimeout: ReturnType<typeof setTimeout> | null = null;
-						const stopProgressPolling = () => {
-							progressPollActive = false;
-							progressPollGeneration++;
-							if (progressPollTimeout) {
-								clearTimeout(progressPollTimeout);
-								progressPollTimeout = null;
-							}
-						};
-						let otherDocsTotal = 0;
-						let otherDocsChecked = 0;
-						for (const doc of documents) {
-							if (doc.filename === docEntry.filename) continue;
-							try {
-								const r = await readDocAndCountTasks(folderPath, doc.filename, sshRemoteId);
-								otherDocsTotal += r.taskCount + r.checkedCount;
-								otherDocsChecked += r.checkedCount;
-							} catch {
-								// Ignore — baseline is best-effort
-							}
-						}
-						const runProgressPoll = async () => {
-							if (!progressPollActive || progressPollInFlight) return;
-							const generationAtStart = progressPollGeneration;
-							progressPollInFlight = true;
-							try {
-								const r = await readDocAndCountTasks(folderPath, docEntry.filename, sshRemoteId);
-								const polledTotal = otherDocsTotal + r.taskCount + r.checkedCount;
-								const polledChecked = otherDocsChecked + r.checkedCount;
-								if (!progressPollActive || generationAtStart !== progressPollGeneration) return;
-								updateBatchStateAndBroadcastRef.current!(sessionId, (prev) => {
-									const prevState = prev[sessionId] || DEFAULT_BATCH_STATE;
-									if (
-										polledChecked === prevState.completedTasksAcrossAllDocs &&
-										polledTotal === prevState.totalTasksAcrossAllDocs
-									) {
-										return prev;
-									}
-									return {
-										...prev,
-										[sessionId]: {
-											...prevState,
-											completedTasksAcrossAllDocs: polledChecked,
-											totalTasksAcrossAllDocs: Math.max(0, polledTotal),
-										},
-									};
-								});
-
-								// Keep the displayed document content fresh during batch runs, even if
-								// file watcher events are coalesced or dropped.
-								const currentSession = sessionsRef.current.find((s) => s.id === sessionId);
-								const selectedDoc = currentSession?.autoRunSelectedFile;
-								if (selectedDoc) {
-									const selectedDocResult = await window.maestro.autorun.readDoc(
-										folderPath,
-										selectedDoc + '.md',
-										sshRemoteId
-									);
-									if (selectedDocResult.success) {
-										if (!progressPollActive || generationAtStart !== progressPollGeneration) return;
-										const nextContent = selectedDocResult.content || '';
-										if (nextContent !== currentSession.autoRunContent) {
-											onUpdateSession(sessionId, {
-												autoRunContent: nextContent,
-												autoRunContentVersion: (currentSession.autoRunContentVersion || 0) + 1,
-											});
-										}
-									}
-								}
-							} catch {
-								// Ignore polling errors — agent may be modifying file
-							} finally {
-								progressPollInFlight = false;
-								if (progressPollActive && generationAtStart === progressPollGeneration) {
-									progressPollTimeout = setTimeout(() => {
-										void runProgressPoll();
-									}, AUTO_RUN_PROGRESS_POLL_INTERVAL_MS);
-								}
-							}
-						};
-						progressPollGeneration++;
-						progressPollTimeout = setTimeout(() => {
-							void runProgressPoll();
-						}, AUTO_RUN_PROGRESS_POLL_INTERVAL_MS);
-
-						try {
-							const taskResult = await documentProcessor.processTask(
-								{
-									folderPath,
-									session,
-									gitBranch,
-									groupName,
-									loopIteration: loopIteration + 1, // 1-indexed
-									effectiveCwd,
-									customPrompt: prompt,
-									sshRemoteId,
-								},
-								effectiveFilename, // Use working copy path for reset-on-completion docs
-								docCheckedCount,
-								remainingTasks,
-								docContent,
-								{
-									onSpawnAgent,
-								}
-							);
-
-							stopProgressPolling();
-
-							// Track agent session IDs
-							if (taskResult.agentSessionId) {
-								agentSessionIds.push(taskResult.agentSessionId);
-							}
-
-							anyTasksProcessedThisIteration = true;
-
-							// Extract results from processTask
-							const {
-								tasksCompletedThisRun,
-								addedUncheckedTasks,
-								newRemainingTasks,
-								documentChanged,
-								newCheckedCount,
-								shortSummary,
-								fullSynopsis,
-								usageStats,
-								contextUsage,
-								elapsedTimeMs,
-								agentSessionId,
-								success,
-								errorKind,
-							} = taskResult;
-
-							// Detect stalling via task-count invariance: if no tasks were checked off AND
-							// the set of tasks didn't change (none added, none removed), the agent made
-							// no real progress this iteration. This ignores prose/addendum churn in the
-							// document — an agent writing "why I did nothing" into the file doesn't
-							// count as progress.
-							const prevCheckedCount = docCheckedCount;
-							const prevUncheckedCount = remainingTasks;
-							const checkedCountChanged = newCheckedCount !== prevCheckedCount;
-							const uncheckedCountChanged = newRemainingTasks !== prevUncheckedCount;
-							const taskSetChanged = checkedCountChanged || uncheckedCountChanged;
-							const prevNoChangeCount = consecutiveNoChangeCount;
-							const beforeLen = docContent?.length ?? 0;
-							const afterLen = taskResult.contentAfterTask?.length ?? 0;
-							const byteDelta = afterLen - beforeLen;
-
-							// Watchdog failures indicate the agent hung or blew past its time budget,
-							// so skip the heuristic and terminate this document immediately.
-							const isWatchdogFailure =
-								errorKind === 'watchdog-stalled' || errorKind === 'watchdog-timeout';
-							if (isWatchdogFailure) {
-								consecutiveNoChangeCount = MAX_CONSECUTIVE_NO_CHANGES;
-							} else if (tasksCompletedThisRun === 0 && !taskSetChanged) {
-								consecutiveNoChangeCount++;
-							} else {
-								consecutiveNoChangeCount = 0;
-							}
-
-							// AUTORUN LOG: stall detection trace — logged every iteration so field
-							// reports can reconstruct why the counter did or did not increment.
-							// `appendOnlyNoProgress` flags the "agent appended explanation text instead
-							// of doing work" pattern: doc bytes grew but the task set is unchanged.
-							window.maestro.logger.autorun(
-								`Stall trace: ${docEntry.filename} iter=${loopIteration + 1} counter=${prevNoChangeCount}->${consecutiveNoChangeCount}/${MAX_CONSECUTIVE_NO_CHANGES}`,
-								session.name,
-								{
-									document: docEntry.filename,
-									loopNumber: loopIteration + 1,
-									documentChanged,
-									tasksCompletedThisRun,
-									prevCheckedCount,
-									newCheckedCount,
-									prevUncheckedCount,
-									newUncheckedCount: newRemainingTasks,
-									checkedCountChanged,
-									uncheckedCountChanged,
-									taskSetChanged,
-									contentLenBefore: beforeLen,
-									contentLenAfter: afterLen,
-									byteDelta,
-									counterBefore: prevNoChangeCount,
-									counterAfter: consecutiveNoChangeCount,
-									maxNoChange: MAX_CONSECUTIVE_NO_CHANGES,
-									appendOnlyNoProgress:
-										documentChanged &&
-										tasksCompletedThisRun === 0 &&
-										!taskSetChanged &&
-										byteDelta > 0,
-									success,
-								}
-							);
-
-							// Update counters
-							docTasksCompleted += tasksCompletedThisRun;
-							totalCompletedTasks += tasksCompletedThisRun;
-							loopTasksCompleted += tasksCompletedThisRun;
-
-							// Record this task in stats database (if stats tracking is active)
-							if (statsAutoRunId && tasksCompletedThisRun > 0) {
-								try {
-									await window.maestro.stats.recordAutoTask({
-										autoRunSessionId: statsAutoRunId,
-										sessionId: sessionId,
-										agentType: session.toolType,
-										taskIndex: totalCompletedTasks - 1, // 0-indexed
-										taskContent: shortSummary || undefined,
-										startTime: Date.now() - elapsedTimeMs,
-										duration: elapsedTimeMs,
-										success: success,
-									});
-								} catch (statsError) {
-									// Don't fail the batch if stats tracking fails
-									logger.warn(
-										'[BatchProcessor] Failed to record task stats:',
-										undefined,
-										statsError
-									);
-								}
-							}
-
-							// Track token usage for loop summary and cumulative totals
-							if (usageStats) {
-								loopTotalInputTokens += usageStats.inputTokens || 0;
-								loopTotalOutputTokens += usageStats.outputTokens || 0;
-								loopTotalCost += usageStats.totalCostUsd || 0;
-								// Also track cumulative totals for final summary
-								totalInputTokens += usageStats.inputTokens || 0;
-								totalOutputTokens += usageStats.outputTokens || 0;
-								totalCost += usageStats.totalCostUsd || 0;
-							}
-
-							// Update Symphony contribution with real-time progress
-							if (session.symphonyMetadata?.isSymphonySession) {
-								window.maestro.symphony
-									.updateStatus({
-										contributionId: session.symphonyMetadata.contributionId,
-										progress: {
-											totalDocuments: documents.length,
-											completedDocuments: docIndex,
-											totalTasks: initialTotalTasks,
-											completedTasks: totalCompletedTasks,
-											currentDocument: docEntry.filename,
-										},
-										tokenUsage: {
-											inputTokens: totalInputTokens,
-											outputTokens: totalOutputTokens,
-											estimatedCost: totalCost,
-										},
-										timeSpent: timeTracking.getElapsedTime(sessionId),
-									})
-									.catch((err: unknown) => {
-										logger.warn(
-											'[BatchProcessor] Failed to update Symphony progress:',
-											undefined,
-											err
-										);
-									});
-							}
-
-							// Track non-reset document completions for loop exit logic
-							// (This tracking is intentionally a no-op for now - kept for future loop mode enhancements)
-							void (!docEntry.resetOnCompletion ? tasksCompletedThisRun : 0);
-
-							// Update progress state for current document
-							if (addedUncheckedTasks > 0) {
-								docTasksTotal += addedUncheckedTasks;
-							}
-
-							// Recount all documents to get accurate total
-							// Tasks in one document can create tasks in other documents,
-							// so delta-based tracking on just the current doc is insufficient
-							let recountedTotal = 0;
-							let recountedChecked = 0;
-							for (const doc of documents) {
-								const { taskCount, checkedCount } = await readDocAndCountTasks(
-									folderPath,
-									doc.filename,
-									sshRemoteId
-								);
-								recountedTotal += taskCount + checkedCount;
-								recountedChecked += checkedCount;
-							}
-
-							updateBatchStateAndBroadcastRef.current!(sessionId, (prev) => {
-								const prevState = prev[sessionId] || DEFAULT_BATCH_STATE;
-								const nextTotalAcrossAllDocs = Math.max(0, recountedTotal);
-								const nextTotalTasks = Math.max(0, recountedTotal);
-
-								return {
-									...prev,
-									[sessionId]: {
-										...prevState,
-										currentDocTasksCompleted: docTasksCompleted,
-										currentDocTasksTotal: docTasksTotal,
-										completedTasksAcrossAllDocs: recountedChecked,
-										totalTasksAcrossAllDocs: nextTotalAcrossAllDocs,
-										// Accumulate actual task duration (most accurate work time tracking)
-										cumulativeTaskTimeMs: (prevState.cumulativeTaskTimeMs || 0) + elapsedTimeMs,
-										// Legacy fields
-										completedTasks: totalCompletedTasks,
-										totalTasks: nextTotalTasks,
-										currentTaskIndex: totalCompletedTasks,
-										sessionIds: [...(prevState?.sessionIds || []), agentSessionId || ''],
-									},
-								};
-							});
-
-							// Add history entry
-							// Use effectiveCwd for projectPath so clicking the session link looks in the right place
-							onAddHistoryEntry({
-								type: 'AUTO',
-								timestamp: Date.now(),
-								summary: shortSummary,
-								fullResponse: fullSynopsis,
-								agentSessionId,
-								projectPath: effectiveCwd,
-								sessionId: sessionId,
-								success,
-								usageStats,
-								contextUsage,
-								elapsedTimeMs,
-							});
-
-							// Speak the synopsis via TTS if audio feedback is enabled
-							// Use refs to get latest setting values (user may toggle mid-run)
-							if (
-								audioFeedbackEnabledRef.current &&
-								audioFeedbackCommandRef.current &&
-								shortSummary
-							) {
-								window.maestro.notification
-									.speak(shortSummary, audioFeedbackCommandRef.current)
-									.catch((err) => {
-										logger.error('[BatchProcessor] Failed to speak synopsis:', undefined, err);
-									});
-							}
-
-							// Check if we've hit the stalling threshold for this document
-							if (consecutiveNoChangeCount >= MAX_CONSECUTIVE_NO_CHANGES) {
-								const stallReason = `${consecutiveNoChangeCount} consecutive runs with no progress`;
-
-								// Track this document as stalled
-								stalledDocuments.set(docEntry.filename, stallReason);
-
-								// AUTORUN LOG: Document stalled
-								window.maestro.logger.autorun(
-									`Document stalled: ${docEntry.filename}`,
-									session.name,
-									{
-										document: docEntry.filename,
-										reason: stallReason,
-										remainingTasks: newRemainingTasks,
-										loopNumber: loopIteration + 1,
-									}
-								);
-
-								// Add a history entry specifically for this stalled document
-								const stallExplanation = [
-									`**Document Stalled: ${docEntry.filename}**`,
-									'',
-									`The AI agent ran ${consecutiveNoChangeCount} times on this document but made no task-level progress:`,
-									`- No tasks were checked off`,
-									`- No tasks were added or removed`,
-									'',
-									`**What this means:**`,
-									`The remaining tasks in this document may be:`,
-									`- Already complete (but not checked off)`,
-									`- Unclear or ambiguous for the AI to act on`,
-									`- Dependent on external factors or manual intervention`,
-									`- Outside the scope of what the AI can accomplish`,
-									'',
-									`**Remaining unchecked tasks:** ${newRemainingTasks}`,
-									'',
-									documents.length > 1
-										? `Skipping to the next document in the playbook...`
-										: `No more documents to process.`,
-								].join('\n');
-
-								onAddHistoryEntry({
-									type: 'AUTO',
-									timestamp: Date.now(),
-									summary: `Document stalled: ${docEntry.filename} (${newRemainingTasks} tasks remaining)`,
-									fullResponse: stallExplanation,
-									projectPath: effectiveCwd,
-									sessionId: sessionId,
-									success: false, // Mark as unsuccessful since we couldn't complete
-								});
-
-								// Skip to the next document instead of breaking the entire batch
-								break; // Break out of the inner while loop for this document
-							}
-
-							docCheckedCount = newCheckedCount;
-							remainingTasks = newRemainingTasks;
-							docContent = taskResult.contentAfterTask;
-						} catch (error) {
-							stopProgressPolling();
-							logger.error(
-								`[BatchProcessor] Error running task in ${docEntry.filename} for session ${sessionId}:`,
-								undefined,
-								error
-							);
-
-							// Check if an error resolution promise was created (e.g., by onAgentError → pauseBatchOnError)
-							// This handles the case where the agent error (e.g., context limit) triggered a pause,
-							// but processTask threw before the next loop iteration could check for it.
-							const postTaskErrorResolution = errorResolutionRefs.current[sessionId];
-							if (postTaskErrorResolution) {
-								const action = await postTaskErrorResolution.promise;
-								delete errorResolutionRefs.current[sessionId];
-
-								if (action === 'abort') {
-									stopRequestedRefs.current[sessionId] = true;
-									break;
-								}
-
-								if (action === 'skip-document') {
-									skipCurrentDocumentAfterError = true;
-									break;
-								}
-
-								// 'resume' — re-read document to get accurate task count before continuing
-								const {
-									taskCount,
-									checkedCount,
-									content: freshContent,
-								} = await readDocAndCountTasks(folderPath, effectiveFilename, sshRemoteId);
-								remainingTasks = taskCount;
-								docCheckedCount = checkedCount;
-								docContent = freshContent;
-								continue;
-							}
-
-							// No error resolution pending — continue to next task on error
-							remainingTasks--;
-						}
-					}
-
-					// Check for stop request before moving to next document
-					if (stopRequestedRefs.current[sessionId]) {
-						break;
-					}
-
-					// Skip document handling if this document stalled (it didn't complete normally)
-					if (stalledDocuments.has(docEntry.filename)) {
-						// Working copy approach: stalled working copy stays in /Runs/ as audit log
-						// Original document is untouched, so nothing to restore
-						workingCopies.delete(docEntry.filename);
-						// Reset consecutive no-change counter for next document
-						consecutiveNoChangeCount = 0;
-						continue;
-					}
-
-					if (skipCurrentDocumentAfterError) {
-						// Working copy approach: errored working copy stays in /Runs/ as audit log
-						// Original document is untouched, so nothing to restore
-						workingCopies.delete(docEntry.filename);
-						continue;
-					}
-
-					// Document complete - for reset-on-completion docs, original is untouched
-					// Working copy in /Runs/ serves as the audit log of this loop's work
-					if (docEntry.resetOnCompletion && docTasksCompleted > 0) {
-						// AUTORUN LOG: Document loop completed
-						window.maestro.logger.autorun(
-							`Document loop completed: ${docEntry.filename}`,
-							session.name,
-							{
-								document: docEntry.filename,
-								workingCopy: workingCopies.get(docEntry.filename),
-								tasksCompleted: docTasksCompleted,
-								loopNumber: loopIteration + 1,
-							}
-						);
-
-						// For loop mode, re-count tasks in the original document for next iteration
-						// (original is unchanged, so it still has all unchecked tasks)
-						if (loopEnabled) {
-							const { taskCount: resetTaskCount } = await readDocAndCountTasks(
-								folderPath,
-								docEntry.filename,
-								sshRemoteId
-							);
-							updateBatchStateAndBroadcastRef.current!(sessionId, (prev) => ({
-								...prev,
-								[sessionId]: {
-									...prev[sessionId],
-									totalTasksAcrossAllDocs: prev[sessionId].totalTasksAcrossAllDocs + resetTaskCount,
-									totalTasks: prev[sessionId].totalTasks + resetTaskCount,
-								},
-							}));
-						}
-
-						// Clear tracking - working copy stays in /Runs/ as audit log
-						workingCopies.delete(docEntry.filename);
-					} else if (docEntry.resetOnCompletion) {
-						// Document had reset enabled but no tasks were completed
-						// Working copy still serves as record of the attempt
-						workingCopies.delete(docEntry.filename);
-					}
-				}
-
-				// Note: We no longer break immediately when a document stalls.
-				// Individual documents that stall are skipped, and we continue processing other documents.
-				// The stalledDocuments map tracks which documents stalled for the final summary.
-
-				// Check if we should continue looping
-				if (!loopEnabled) {
-					// No loop mode - we're done after one pass
-					// AUTORUN LOG: Exit (non-loop mode)
-					window.maestro.logger.autorun(`Auto Run completed (single pass)`, session.name, {
-						reason: 'Single pass completed',
-						totalTasksCompleted: totalCompletedTasks,
-						loopsCompleted: 1,
-					});
-					break;
-				}
-
-				// Check if we've hit the max loop limit
-				if (maxLoops !== null && maxLoops !== undefined && loopIteration + 1 >= maxLoops) {
-					addFinalLoopSummary(`Reached max loop limit (${maxLoops})`);
-					break;
-				}
-
-				// Check for stop request after full pass
-				if (stopRequestedRefs.current[sessionId]) {
-					addFinalLoopSummary('Stopped by user');
-					break;
-				}
-
-				// Safety check: if we didn't process ANY tasks this iteration, exit to avoid infinite loop
-				if (!anyTasksProcessedThisIteration) {
-					addFinalLoopSummary('No tasks processed this iteration');
-					break;
-				}
-
-				// Loop mode: check if we should continue looping
-				// Check if there are any non-reset documents in the playbook
-				const hasAnyNonResetDocs = documents.some((doc) => !doc.resetOnCompletion);
-
-				if (hasAnyNonResetDocs) {
-					// If we have non-reset docs, only continue if they have remaining tasks
-					let anyNonResetDocsHaveTasks = false;
-					for (const doc of documents) {
-						if (doc.resetOnCompletion) continue;
-
-						const { taskCount } = await readDocAndCountTasks(folderPath, doc.filename, sshRemoteId);
-						if (taskCount > 0) {
-							anyNonResetDocsHaveTasks = true;
-							break;
-						}
-					}
-
-					if (!anyNonResetDocsHaveTasks) {
-						addFinalLoopSummary('All tasks completed');
-						break;
-					}
-				}
-				// If all documents are reset docs, we continue looping (maxLoops check above will stop us)
-
-				// Re-scan all documents to get fresh task counts for next loop (tasks may have been added/removed)
-				let newTotalTasks = 0;
-				for (const doc of documents) {
-					const { taskCount } = await readDocAndCountTasks(folderPath, doc.filename, sshRemoteId);
-					newTotalTasks += taskCount;
-				}
-
-				// Capture completed-loop metrics before resetting counters
-				const completedLoopNumber = loopIteration + 1;
-				const completedLoopTasks = loopTasksCompleted;
-
-				// Calculate loop elapsed time
-				const loopElapsedMs = Date.now() - loopStartTime;
-
-				// Add loop summary history entry
-				const loopSummary = `Loop ${completedLoopNumber} completed: ${completedLoopTasks} task${completedLoopTasks !== 1 ? 's' : ''} accomplished`;
-				const loopDetails = [
-					`**Loop ${completedLoopNumber} Summary**`,
-					'',
-					`- **Tasks Accomplished:** ${completedLoopTasks}`,
-					`- **Duration:** ${formatElapsedTime(loopElapsedMs)}`,
-					loopTotalInputTokens > 0 || loopTotalOutputTokens > 0
-						? `- **Tokens:** ${(loopTotalInputTokens + loopTotalOutputTokens).toLocaleString()} (${loopTotalInputTokens.toLocaleString()} in / ${loopTotalOutputTokens.toLocaleString()} out)`
-						: '',
-					loopTotalCost > 0 ? `- **Cost:** $${loopTotalCost.toFixed(4)}` : '',
-					`- **Tasks Discovered for Next Loop:** ${newTotalTasks}`,
-				]
-					.filter((line) => line !== '')
-					.join('\n');
-
-				onAddHistoryEntry({
-					type: 'AUTO',
-					timestamp: Date.now(),
-					summary: loopSummary,
-					fullResponse: loopDetails,
-					projectPath: session.cwd,
-					sessionId: sessionId,
-					success: true,
-					elapsedTimeMs: loopElapsedMs,
-					usageStats:
-						loopTotalInputTokens > 0 || loopTotalOutputTokens > 0
-							? {
-									inputTokens: loopTotalInputTokens,
-									outputTokens: loopTotalOutputTokens,
-									cacheReadInputTokens: 0,
-									cacheCreationInputTokens: 0,
-									totalCostUsd: loopTotalCost,
-									contextWindow: 0,
-								}
-							: undefined,
-				});
-
-				// Reset per-loop tracking for next iteration
-				loopStartTime = Date.now();
-				loopTasksCompleted = 0;
-				loopTotalInputTokens = 0;
-				loopTotalOutputTokens = 0;
-				loopTotalCost = 0;
-
-				// AUTORUN LOG: Loop completion
-				window.maestro.logger.autorun(`Loop ${completedLoopNumber} completed`, session.name, {
-					loopNumber: completedLoopNumber,
-					tasksCompleted: completedLoopTasks,
-					tasksForNextLoop: newTotalTasks,
-				});
-
-				// Continue looping
-				loopIteration++;
-
-				updateBatchStateAndBroadcastRef.current!(sessionId, (prev) => ({
-					...prev,
-					[sessionId]: {
-						...prev[sessionId],
-						loopIteration,
-						totalTasksAcrossAllDocs: newTotalTasks + prev[sessionId].completedTasksAcrossAllDocs,
-						totalTasks: newTotalTasks + prev[sessionId].completedTasks,
-					},
-				}));
-			}
-
-			// Working copy approach: no cleanup needed
-			// - Original documents are never modified
-			// - Working copies in /Runs/ serve as audit logs and are kept
-			// - User can delete them manually if desired
-
-			// Create PR if worktree was used, PR creation is enabled, and not stopped
-			const wasStopped = stopRequestedRefs.current[sessionId] || false;
-			const sessionName = session.name || session.cwd.split('/').pop() || 'Unknown';
-			if (
-				worktreeActive &&
-				worktree?.createPROnCompletion &&
-				!wasStopped &&
-				totalCompletedTasks > 0 &&
-				worktreePath
-			) {
-				// For worktree-dispatched runs, the main repo is the parent session's cwd
-				const mainRepoCwd = config.worktreeTarget
-					? sessionsRef.current.find((s) => s.id === session.parentSessionId)?.cwd || session.cwd
-					: session.cwd;
-
-				const prResult = await worktreeManager.createPR({
-					worktreePath,
-					mainRepoCwd,
-					worktree,
-					documents,
-					totalCompletedTasks,
-				});
-
-				if (onPRResult) {
-					onPRResult({
-						sessionId,
-						sessionName,
-						success: prResult.success,
-						prUrl: prResult.prUrl,
-						error: prResult.error,
-					});
-				}
-
-				// Record PR result in history so it's visible in the worktree agent's history panel
-				onAddHistoryEntry({
-					type: 'AUTO',
-					timestamp: Date.now(),
-					summary: prResult.success
-						? `PR created: ${prResult.prUrl}`
-						: `PR creation failed: ${prResult.error || 'Unknown error'}`,
-					fullResponse: prResult.success
-						? `**Pull Request Created**\n\n- **URL:** ${prResult.prUrl}\n- **Branch:** \`${worktreeBranch}\`\n- **Target:** \`${prResult.targetBranch || 'unknown'}\`\n- **Tasks Completed:** ${totalCompletedTasks}`
-						: `**Pull Request Creation Failed**\n\n- **Error:** ${prResult.error || 'Unknown error'}\n- **Branch:** \`${worktreeBranch}\`\n- **Target:** \`${prResult.targetBranch || 'unknown'}\``,
-					projectPath: worktreePath,
-					sessionId,
-					success: prResult.success,
-				});
-			}
-
-			// Add final Auto Run summary entry
-			// Calculate visibility-aware elapsed time using the extracted time tracking hook
-			// (excludes time when laptop was sleeping/suspended)
-			const totalElapsedMs = timeTracking.getElapsedTime(sessionId);
-			const loopsCompleted = loopEnabled ? loopIteration + 1 : 1;
-
-			// Determine status based on stalled documents and completion
-			const stalledCount = stalledDocuments.size;
-			const allDocsStalled = stalledCount === documents.length;
-			const someDocsStalled = stalledCount > 0 && stalledCount < documents.length;
-			const statusText = wasStopped
-				? 'stopped'
-				: allDocsStalled
-					? 'stalled'
-					: someDocsStalled
-						? 'completed with stalls'
-						: 'completed';
-
-			// Calculate achievement progress for the summary
-			// Note: We use the stats BEFORE this run is recorded (the parent will call recordAutoRunComplete after)
-			// So we need to add totalElapsedMs to get the projected cumulative time
-			const projectedCumulativeTime = (autoRunStats?.cumulativeTimeMs || 0) + totalElapsedMs;
-			const currentBadge = getBadgeForTime(projectedCumulativeTime);
-			const nextBadge = getNextBadge(currentBadge);
-			const levelProgressText = nextBadge
-				? `Level ${currentBadge?.level || 0} → ${nextBadge.level}: ${formatTimeRemaining(projectedCumulativeTime, nextBadge)}`
-				: currentBadge
-					? `Level ${currentBadge.level} (${currentBadge.name}) - Maximum level achieved!`
-					: 'Level 0 → 1: ' + formatTimeRemaining(0, getBadgeForTime(0));
-
-			// Build summary with stall info if applicable
-			const stalledSuffix = stalledCount > 0 ? ` (${stalledCount} stalled)` : '';
-			const finalSummary = `Auto Run ${statusText}: ${totalCompletedTasks} task${totalCompletedTasks !== 1 ? 's' : ''} in ${formatElapsedTime(totalElapsedMs)}${stalledSuffix}`;
-
-			// Build status message with detailed info
-			let statusMessage: string;
-			if (wasStopped) {
-				statusMessage = 'Stopped by user';
-			} else if (allDocsStalled) {
-				statusMessage = `Stalled - All ${stalledCount} document(s) stopped making progress`;
-			} else if (someDocsStalled) {
-				statusMessage = `Completed with ${stalledCount} stalled document(s)`;
-			} else {
-				statusMessage = 'Completed';
-			}
-
-			// Build stalled documents section if any documents stalled
-			const stalledDocsSection: string[] = [];
-			if (stalledCount > 0) {
-				stalledDocsSection.push('');
-				stalledDocsSection.push('**Stalled Documents**');
-				stalledDocsSection.push('');
-				stalledDocsSection.push(
-					'The following documents stopped making progress after multiple attempts:'
-				);
-				for (const [docName, reason] of stalledDocuments) {
-					stalledDocsSection.push(`- **${docName}**: ${reason}`);
-				}
-				stalledDocsSection.push('');
-				stalledDocsSection.push(
-					'*Tasks in stalled documents may need manual review or clarification.*'
-				);
-			}
-
-			const finalDetails = [
-				`**Auto Run Summary**`,
-				'',
-				`- **Status:** ${statusMessage}`,
-				`- **Tasks Completed:** ${totalCompletedTasks}`,
-				`- **Total Duration:** ${formatElapsedTime(totalElapsedMs)}`,
-				loopEnabled ? `- **Loops Completed:** ${loopsCompleted}` : '',
-				totalInputTokens > 0 || totalOutputTokens > 0
-					? `- **Total Tokens:** ${(totalInputTokens + totalOutputTokens).toLocaleString()} (${totalInputTokens.toLocaleString()} in / ${totalOutputTokens.toLocaleString()} out)`
-					: '',
-				totalCost > 0 ? `- **Total Cost:** $${totalCost.toFixed(4)}` : '',
-				'',
-				`- **Documents:** ${documents.map((d) => d.filename).join(', ')}`,
-				...stalledDocsSection,
-				'',
-				`**Achievement Progress**`,
-				`- ${levelProgressText}`,
-			]
-				.filter((line) => line !== '')
-				.join('\n');
-
-			// Success is true if not stopped and at least some documents completed without stalling
-			const isSuccess = !wasStopped && !allDocsStalled;
-
-			// Claim the flush state: if killBatchRun already flushed, skip the history + stats
-			// writes here to avoid clobbering recorded duration with a stale/zero value.
-			const alreadyFlushed = !autoRunFlushStateRefs.current[sessionId];
-			delete autoRunFlushStateRefs.current[sessionId];
-
-			if (!alreadyFlushed) {
-				try {
-					await onAddHistoryEntry({
-						type: 'AUTO',
-						timestamp: Date.now(),
-						summary: finalSummary,
-						fullResponse: finalDetails,
-						projectPath: session.cwd,
-						sessionId, // Include sessionId so the summary appears in session's history
-						success: isSuccess,
-						elapsedTimeMs: totalElapsedMs,
-						usageStats:
-							totalInputTokens > 0 || totalOutputTokens > 0
-								? {
-										inputTokens: totalInputTokens,
-										outputTokens: totalOutputTokens,
-										cacheReadInputTokens: 0,
-										cacheCreationInputTokens: 0,
-										totalCostUsd: totalCost,
-										contextWindow: 0,
-									}
-								: undefined,
-						achievementAction: 'openAbout', // Enable clickable link to achievements panel
-					});
-				} catch {
-					// Ignore history errors
-				}
-
-				// End stats tracking for this Auto Run session
-				if (statsAutoRunId) {
-					try {
-						await window.maestro.stats.endAutoRun(
-							statsAutoRunId,
-							totalElapsedMs,
-							totalCompletedTasks
-						);
-					} catch (statsError) {
-						// Don't fail cleanup if stats tracking fails
-						logger.warn('[BatchProcessor] Failed to end stats tracking:', undefined, statsError);
-					}
-				}
-			}
-
-			// Critical: Always flush debounced updates and dispatch COMPLETE_BATCH to clean up state.
-			// These operations are safe regardless of mount state - React handles reducer dispatches gracefully,
-			// and broadcasts are external calls that don't affect React state.
-			flushDebouncedUpdate(sessionId);
-
-			// Reset state for this session using COMPLETE_BATCH action
-			// (not updateBatchStateAndBroadcast which only supports UPDATE_PROGRESS)
-			dispatch({
-				type: 'COMPLETE_BATCH',
-				sessionId,
-				finalSessionIds: agentSessionIds,
-			});
-			// Broadcast state change to web clients
-			broadcastAutoRunState(sessionId, null);
-
-			// Call completion callback if provided (only if still mounted to avoid warnings).
-			// Skip when alreadyFlushed: killBatchRun owns the onComplete call in that case
-			// (it captured the elapsed time before stopTracking zeroed it). Invoking here
-			// would double-fire the toast and submit elapsedTimeMs:0 to the leaderboard.
-			if (!alreadyFlushed && isMountedRef.current && onComplete) {
-				onComplete({
-					sessionId,
-					sessionName: session.name || session.cwd.split('/').pop() || 'Unknown',
-					completedTasks: totalCompletedTasks,
-					totalTasks: initialTotalTasks,
-					wasStopped,
-					elapsedTimeMs: totalElapsedMs,
-					inputTokens: totalInputTokens,
-					outputTokens: totalOutputTokens,
-					totalCostUsd: totalCost,
-					documentsProcessed: documents.length,
-				});
-			}
-
-			// Process any queued items that were waiting during batch run
-			// This ensures pending user messages are processed after Auto Run ends
-			if (isMountedRef.current && onProcessQueueAfterCompletion) {
-				// Use setTimeout to let state updates settle before processing queue
-				setTimeout(() => {
-					onProcessQueueAfterCompletion(sessionId);
-				}, 0);
-			}
-
-			// Clean up time tracking, error resolution, and stop request flag
-			// Clearing stopRequestedRefs here (not just at start) ensures proper cleanup
-			// regardless of how the batch ended (normal completion, stopped, or error)
-			// Note: These cleanup operations are safe even after unmount (they only affect refs)
-			timeTracking.stopTracking(sessionId);
-			delete errorResolutionRefs.current[sessionId];
-			delete stopRequestedRefs.current[sessionId];
-
-			// Allow system to sleep now that Auto Run is complete
-			window.maestro.power.removeReason(`autorun:${sessionId}`);
-			// Note: updateBatchStateAndBroadcast is accessed via ref to avoid stale closure in long-running async
-			// flushDebouncedUpdate is stable (empty deps in useSessionDebounce) so adding it doesn't cause re-renders
-		},
-		// Note: audioFeedbackEnabled/audioFeedbackCommand removed from deps - we use refs
-		// to allow mid-run setting changes to take effect immediately
-		[
-			onUpdateSession,
-			onSpawnAgent,
-			onAddHistoryEntry,
-			onComplete,
-			onPRResult,
-			timeTracking,
-			onProcessQueueAfterCompletion,
-			flushDebouncedUpdate,
-		]
-	);
-
-	/**
-	 * Request to stop the batch run for a specific session after current task completes
-	 * Note: No isMountedRef check here - stop requests should always be honored.
-	 * All operations are safe: ref updates, reducer dispatch (React handles gracefully), and broadcasts.
-	 */
-	const stopBatchRun = useCallback(
-		(sessionId: string) => {
-			stopRequestedRefs.current[sessionId] = true;
-			const errorResolution = errorResolutionRefs.current[sessionId];
-			if (errorResolution) {
-				errorResolution.resolve('abort');
-				delete errorResolutionRefs.current[sessionId];
-			}
-			// Use SET_STOPPING action directly (not updateBatchStateAndBroadcast which only supports UPDATE_PROGRESS)
-			dispatch({ type: 'SET_STOPPING', sessionId });
-			// Broadcast state change
-			const newState = useBatchStore.getState().batchRunStates[sessionId];
-			if (newState) {
-				broadcastAutoRunState(sessionId, { ...newState, isStopping: true });
-			}
-		},
-		[broadcastAutoRunState]
-	);
-
-	/**
-	 * Force kill the running process and immediately end the batch run.
-	 * Unlike stopBatchRun (which waits for the current task to complete),
-	 * this terminates the agent process immediately and resets all batch state.
-	 */
-	const killBatchRun = useCallback(
-		async (sessionId: string) => {
-			console.assert(
-				!sessionId.includes('-batch-'),
-				'[BatchProcessor:killBatchRun] sessionId must not contain "-batch-"'
-			);
-
-			// 0. Flush Auto Run stats + history BEFORE we tear down timeTracking below.
-			//    stopTracking() deletes the tracker, so elapsed time must be captured now.
-			//    Atomically claiming the ref ensures the loop's normal cleanup won't double-write.
-			const flushState = autoRunFlushStateRefs.current[sessionId];
-			delete autoRunFlushStateRefs.current[sessionId];
-			if (flushState) {
-				const elapsedMs = timeTracking.getElapsedTime(sessionId);
-				const completedTasks = flushState.getCompletedTasks();
-				if (flushState.statsAutoRunId) {
-					try {
-						await window.maestro.stats.endAutoRun(
-							flushState.statsAutoRunId,
-							elapsedMs,
-							completedTasks
-						);
-					} catch (statsError) {
-						logger.warn(
-							'[BatchProcessor:killBatchRun] Failed to end stats tracking:',
-							undefined,
-							statsError
-						);
-					}
-				}
-				try {
-					await onAddHistoryEntry({
-						type: 'AUTO',
-						timestamp: Date.now(),
-						summary: `Auto Run killed: ${completedTasks} task${completedTasks !== 1 ? 's' : ''} in ${formatElapsedTime(elapsedMs)}`,
-						fullResponse: [
-							'**Auto Run Summary**',
-							'',
-							'- **Status:** Killed by user',
-							`- **Tasks Completed:** ${completedTasks}`,
-							`- **Total Duration:** ${formatElapsedTime(elapsedMs)}`,
-						].join('\n'),
-						projectPath: flushState.projectPath,
-						sessionId,
-						success: false,
-						elapsedTimeMs: elapsedMs,
-					});
-				} catch (historyError) {
-					logger.warn(
-						'[BatchProcessor:killBatchRun] Failed to add history entry:',
-						undefined,
-						historyError
-					);
-				}
-
-				// Fire onComplete here so the kill path records local stats and submits to
-				// the leaderboard. The natural-loop cleanup is unreliable for this: it calls
-				// timeTracking.stopTracking before reading getElapsedTime, so it would invoke
-				// onComplete with elapsedTimeMs:0, which the handler gates out.
-				if (isMountedRef.current && onComplete) {
-					onComplete({
-						sessionId,
-						sessionName: flushState.sessionName,
-						completedTasks,
-						totalTasks: flushState.getTotalTasks(),
-						wasStopped: true,
-						elapsedTimeMs: elapsedMs,
-						inputTokens: flushState.getInputTokens(),
-						outputTokens: flushState.getOutputTokens(),
-						totalCostUsd: flushState.getTotalCost(),
-						documentsProcessed: flushState.getDocumentsProcessed(),
-					});
-				}
-			}
-
-			// 1. Kill all active batch processes for this session and wait for termination before cleanup.
-			// Batch process session IDs are generated as: `${sessionId}-batch-${timestamp}`.
-			try {
-				const activeProcesses = await window.maestro.process.getActiveProcesses();
-				const batchProcessIds = activeProcesses
-					.filter(
-						// Intentional scope: kill the root session process and any descendant
-						// auto-run task processes prefixed with `${sessionId}-batch-`.
-						(proc) =>
-							proc.sessionId === sessionId || proc.sessionId.startsWith(`${sessionId}-batch-`)
-					)
-					.map((proc) => proc.sessionId);
-
-				// Fallback to legacy direct ID in case process listing is stale.
-				if (batchProcessIds.length === 0) {
-					batchProcessIds.push(sessionId);
-				}
-
-				await Promise.allSettled(batchProcessIds.map((id) => window.maestro.process.kill(id)));
-			} catch (error) {
-				logger.error('[BatchProcessor:killBatchRun] Failed to kill process:', undefined, error);
-			}
-
-			// 2. Set stop flag so the processing loop exits if it's still running
-			stopRequestedRefs.current[sessionId] = true;
-
-			// 3. Resolve any pending error state
-			const errorResolution = errorResolutionRefs.current[sessionId];
-			if (errorResolution) {
-				errorResolution.resolve('abort');
-				delete errorResolutionRefs.current[sessionId];
-			}
-
-			// 4. Flush any debounced state updates
-			flushDebouncedUpdate(sessionId);
-
-			// 5. Immediately reset batch state
-			dispatch({
-				type: 'COMPLETE_BATCH',
-				sessionId,
-				finalSessionIds: [],
-			});
-
-			// 6. Broadcast cleared state to web clients
-			broadcastAutoRunState(sessionId, null);
-
-			// 7. Clean up tracking
-			timeTracking.stopTracking(sessionId);
-			// Intentionally do NOT delete stopRequestedRefs[sessionId] here. The async
-			// processing loop is still mid-iteration (its in-flight processTask is about
-			// to reject because we just killed the process). It re-checks the stop flag
-			// at the next iteration boundary (lines ~880, ~891, ~997). Deleting the flag
-			// synchronously here turns it back into `undefined` (falsy) before the loop
-			// observes it, so the loop spawns a fresh agent for the next task and the
-			// "kill" effectively does nothing. The loop's natural-completion cleanup at
-			// the end of startBatchRun handles the delete once it has actually exited.
-			//
-			// 8. Allow system to sleep
-			window.maestro.power.removeReason(`autorun:${sessionId}`);
-		},
-		[broadcastAutoRunState, flushDebouncedUpdate, timeTracking, onAddHistoryEntry, onComplete]
-	);
-
-	/**
-	 * Pause the batch run due to an agent error (Phase 5.10)
-	 * Called externally when agent error is detected
-	 */
-	const pauseBatchOnError = useCallback(
-		(sessionId: string, error: AgentError, documentIndex: number, taskDescription?: string) => {
-			if (!isMountedRef.current) return;
-
-			// Log detailed error to system logs with full context
-			window.maestro.logger.autorun(
-				`Auto Run paused due to ${error.type}: ${error.message}`,
-				sessionId,
-				{
-					errorType: error.type,
-					errorMessage: error.message,
-					recoverable: error.recoverable,
-					documentIndex,
-					taskDescription,
-					rawError: error.raw,
-				}
-			);
-
-			// Use SET_ERROR action directly (not updateBatchStateAndBroadcast which only supports UPDATE_PROGRESS)
-			dispatch({
-				type: 'SET_ERROR',
-				sessionId,
-				payload: { error, documentIndex, taskDescription },
-			});
-			// Broadcast state change
-			const currentState = useBatchStore.getState().batchRunStates[sessionId];
-			if (currentState) {
-				broadcastAutoRunState(sessionId, {
-					...currentState,
-					error,
-					errorPaused: true,
-					errorDocumentIndex: documentIndex,
-					errorTaskDescription: taskDescription,
-				});
-			}
-
-			if (!errorResolutionRefs.current[sessionId]) {
-				let resolvePromise: ((action: ErrorResolutionAction) => void) | undefined;
-				const promise = new Promise<ErrorResolutionAction>((resolve) => {
-					resolvePromise = resolve;
-				});
-				errorResolutionRefs.current[sessionId] = {
-					promise,
-					resolve: resolvePromise as (action: ErrorResolutionAction) => void,
-				};
-			}
-		},
-		[broadcastAutoRunState]
-	);
-
-	/**
-	 * Skip the current document that caused an error and continue with the next one (Phase 5.10)
-	 */
-	const skipCurrentDocument = useCallback(
-		(sessionId: string) => {
-			if (!isMountedRef.current) return;
-
-			window.maestro.logger.autorun(`Skipping document after error`, sessionId, {});
-
-			// Use CLEAR_ERROR action directly (not updateBatchStateAndBroadcast which only supports UPDATE_PROGRESS)
-			dispatch({ type: 'CLEAR_ERROR', sessionId });
-			// Broadcast state change
-			const currentState = useBatchStore.getState().batchRunStates[sessionId];
-			if (currentState) {
-				broadcastAutoRunState(sessionId, {
-					...currentState,
-					error: undefined,
-					errorPaused: false,
-					errorDocumentIndex: undefined,
-					errorTaskDescription: undefined,
-				});
-			}
-
-			const errorResolution = errorResolutionRefs.current[sessionId];
-			if (errorResolution) {
-				errorResolution.resolve('skip-document');
-				delete errorResolutionRefs.current[sessionId];
-			}
-
-			// Signal to skip the current document in the processing loop
-		},
-		[broadcastAutoRunState]
-	);
-
-	/**
-	 * Resume the batch run after an error has been resolved (Phase 5.10)
-	 * This clears the error state and allows the batch to continue
-	 */
-	const resumeAfterError = useCallback(
-		(sessionId: string) => {
-			if (!isMountedRef.current) return;
-
-			window.maestro.logger.autorun(`Resuming Auto Run after error resolution`, sessionId, {});
-
-			// Use CLEAR_ERROR action directly (not updateBatchStateAndBroadcast which only supports UPDATE_PROGRESS)
-			dispatch({ type: 'CLEAR_ERROR', sessionId });
-			// Broadcast state change
-			const currentState = useBatchStore.getState().batchRunStates[sessionId];
-			if (currentState) {
-				broadcastAutoRunState(sessionId, {
-					...currentState,
-					error: undefined,
-					errorPaused: false,
-					errorDocumentIndex: undefined,
-					errorTaskDescription: undefined,
-				});
-			}
-
-			const errorResolution = errorResolutionRefs.current[sessionId];
-			if (errorResolution) {
-				errorResolution.resolve('resume');
-				delete errorResolutionRefs.current[sessionId];
-			}
-		},
-		[broadcastAutoRunState]
-	);
-
-	/**
-	 * Abort the batch run completely due to an unrecoverable error (Phase 5.10)
-	 */
-	const abortBatchOnError = useCallback(
-		(sessionId: string) => {
-			if (!isMountedRef.current) return;
-
-			window.maestro.logger.autorun(`Auto Run aborted due to error`, sessionId, {});
-
-			// Request stop and clear error state
-			stopRequestedRefs.current[sessionId] = true;
-			const errorResolution = errorResolutionRefs.current[sessionId];
-			if (errorResolution) {
-				errorResolution.resolve('abort');
-				delete errorResolutionRefs.current[sessionId];
-			}
-
-			// Use SET_STOPPING action directly (not updateBatchStateAndBroadcast which only
-			// supports UPDATE_PROGRESS and silently drops errorPaused/isStopping/error fields).
-			// SET_STOPPING from PAUSED_ERROR state already clears all error fields.
-			dispatch({ type: 'SET_STOPPING', sessionId });
-			// Broadcast state change
-			const currentState = useBatchStore.getState().batchRunStates[sessionId];
-			if (currentState) {
-				broadcastAutoRunState(sessionId, currentState);
-			}
-		},
-		[broadcastAutoRunState]
-	);
+	// Auto Run orchestrator (the main `startBatchRun` callback)
+	const { startBatchRun } = useBatchRunner({
+		sessionsRef,
+		audioFeedbackEnabledRef,
+		audioFeedbackCommandRef,
+		autoRunFlushStateRefs,
+		errorResolutionRefs,
+		stopRequestedRefs,
+		isMountedRef,
+		updateBatchStateAndBroadcastRef,
+		broadcastAutoRunState,
+		flushDebouncedUpdate,
+		dispatch,
+		timeTracking,
+		worktreeManager,
+		documentProcessor,
+		groups,
+		autoRunStats,
+		onUpdateSession,
+		onSpawnAgent,
+		onAddHistoryEntry,
+		onComplete,
+		onPRResult,
+		onProcessQueueAfterCompletion,
+	});
 
 	return {
 		batchRunStates,
